chore: rename personal wrapper
